+++ Editor/Playa/Utils/SaintsEditorUtils.cs
@@ -153,213 +153,213 @@ namespace SaintsField.Editor.Playa.Utils
             return (showIfResult, disableIfResult);
         }
 
-        public static IReadOnlyList<SerializedInfo> GetSaintsSerialized(Type monoClass)
-        {
-            List<Type> types = ReflectUtils.GetSelfAndBaseTypesFromType(monoClass);
+        // public static IReadOnlyList<SerializedInfo> GetSaintsSerialized(Type monoClass)
+        // {
+        //     List<Type> types = ReflectUtils.GetSelfAndBaseTypesFromType(monoClass);
+        //
+        //     List<SerializedInfo> results = new List<SerializedInfo>();
+        //
+        //     // Debug.Log($"main={monoClass}; types={string.Join(", ", types)}");
+        //     // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
+        //     foreach (Type type in types)
+        //     {
+        //         // Debug.Log(type);
+        //         MemberInfo[] memberLis = type
+        //             .GetMembers(BindingFlags.Instance | BindingFlags.NonPublic |
+        //                         BindingFlags.Public | BindingFlags.DeclaredOnly);
+        //
+        //         foreach (MemberInfo memberInfo in memberLis)
+        //         {
+        //             if (memberInfo.Name.StartsWith("<") && memberInfo.Name.EndsWith(">k__BackingField"))
+        //             {
+        //                 continue;
+        //             }
+        //
+        //             SerializedInfo result;
+        //             switch (memberInfo)
+        //             {
+        //                 case FieldInfo fi:
+        //                 {
+        //                     result = GetSerializedInfo(fi, false, fi.FieldType);
+        //                     // Debug.Log($"{fi.FieldType.IsArray}.{fi.FieldType}");
+        //                 }
+        //                     break;
+        //                 case PropertyInfo pi:
+        //                 {
+        //                     result = GetSerializedInfo(pi, true, pi.PropertyType);
+        //                     // Debug.Log($"{pi.PropertyType.IsArray}.{pi.PropertyType.IsArray}");
+        //                 }
+        //                     break;
+        //                 default:
+        //                     continue;
+        //             }
+        //
+        //             if (result != null)
+        //             {
+        //                 results.Add(result);
+        //             }
+        //         }
+        //     }
+        //
+        //     // return results.Count > 0? (mainFilePath, results): default;
+        //     return results;
+        // }
 
-            List<SerializedInfo> results = new List<SerializedInfo>();
+//         private static SerializedInfo GetSerializedInfo(MemberInfo memberInfo, bool isProperty, Type targetType)
+//         {
+//             if (targetType == typeof(SaintsSerializedProperty))
+//             {
+//                 return null;
+//             }
+//
+//             Attribute[] attributes = ReflectCache.GetCustomAttributes<Attribute>(memberInfo);
+//             bool hasNonSer = false;
+//             bool hasSaintsSer = false;
+//             foreach (Attribute attribute in attributes)
+//             {
+//                 switch (attribute)
+//                 {
+//                     case NonSerializedAttribute _:
+//                         hasNonSer = true;
+//                         break;
+//                     case SaintsSerializedAttribute _:
+//                         hasSaintsSer = true;
+//                         break;
+//                 }
+//             }
+//
+//             if (hasNonSer && !hasSaintsSer)
+//             {
+//                 return null;
+//             }
+//
+//             SaintsTargetCollection targetCollection = SaintsTargetCollection.FieldOrProperty;
+//             Type elementType = null;
+//             if (targetType.IsArray)
+//             {
+//                 targetCollection = SaintsTargetCollection.Array;
+//                 elementType = targetType.GetElementType();
+//             }
+//             else
+//             {
+//                 Type lisType = GetList(targetType);
+//                 // ReSharper disable once InvertIf
+//                 if (lisType != null)
+//                 {
+//                     targetCollection = SaintsTargetCollection.List;
+//                     elementType = lisType.GetGenericArguments()[0];
+//                 }
+//             }
+//
+//             Type checkingType = targetCollection == SaintsTargetCollection.FieldOrProperty
+//                 ? targetType
+//                 : elementType;
+//
+//             Debug.Assert(checkingType != null, targetType);
+//
+//             // ReSharper disable once PossibleNullReferenceException
+//             bool checkingIsEnum = checkingType.IsEnum;
+//
+//             // ReSharper disable once PossibleNullReferenceException
+//             string fillTypeNameWithNameSpace = checkingType.FullName.Replace('+', '.');
+//
+//             if (!checkingIsEnum && IsGeneralClassOrStruct(checkingType))
+//             {
+//                 if (!checkingType.IsDefined(typeof(SerializableAttribute)))
+//                 {
+//                     return null;
+//                 }
+//
+//                 SerializedInfo r = new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection, SaintsPropertyType.ClassOrStruct);
+//
+//                 // class/struct check
+//                 foreach (MemberInfo subMember in checkingType
+//                              .GetMembers(BindingFlags.Instance | BindingFlags.NonPublic |
+//                                          BindingFlags.Public | BindingFlags.DeclaredOnly))
+//                 {
+//                     SerializedInfo subSerInfo;
+//                     switch (subMember)
+//                     {
+//                         case FieldInfo fi:
+//                             subSerInfo = GetSerializedInfo(fi, false, fi.FieldType);
+//                             break;
+//                         case PropertyInfo pi:
+//                             subSerInfo = GetSerializedInfo(pi, true, pi.PropertyType);
+//                             break;
+//                         default:
+//                             continue;
+//                     }
+//
+//                     if (subSerInfo != null)
+//                     {
+//                         r.SubFields.Add(subSerInfo);
+//                     }
+//                 }
+//
+//                 return r.SubFields.Count > 0 ? r : null;
+//             }
+//
+//             if (!checkingIsEnum)
+//             {
+//                 return null;
+//             }
+//
+//             Type underType = checkingType.GetEnumUnderlyingType();
+//             bool isLong = underType == typeof(long);
+//             if (isLong)
+//             {
+//                 return new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection,
+//                     SaintsPropertyType.EnumLong);
+//             }
+//
+// #if UNITY_2022_1_OR_NEWER
+//             bool isULong = underType == typeof(ulong);
+//             if (isULong)
+//             {
+//                 return new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection,
+//                     SaintsPropertyType.EnumULong);
+//             }
+// #endif
+//
+//             return null;
+//         }
 
-            // Debug.Log($"main={monoClass}; types={string.Join(", ", types)}");
-            // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
-            foreach (Type type in types)
-            {
-                // Debug.Log(type);
-                MemberInfo[] memberLis = type
-                    .GetMembers(BindingFlags.Instance | BindingFlags.NonPublic |
-                                BindingFlags.Public | BindingFlags.DeclaredOnly);
-
-                foreach (MemberInfo memberInfo in memberLis)
-                {
-                    if (memberInfo.Name.StartsWith("<") && memberInfo.Name.EndsWith(">k__BackingField"))
-                    {
-                        continue;
-                    }
-
-                    SerializedInfo result;
-                    switch (memberInfo)
-                    {
-                        case FieldInfo fi:
-                        {
-                            result = GetSerializedInfo(fi, false, fi.FieldType);
-                            // Debug.Log($"{fi.FieldType.IsArray}.{fi.FieldType}");
-                        }
-                            break;
-                        case PropertyInfo pi:
-                        {
-                            result = GetSerializedInfo(pi, true, pi.PropertyType);
-                            // Debug.Log($"{pi.PropertyType.IsArray}.{pi.PropertyType.IsArray}");
-                        }
-                            break;
-                        default:
-                            continue;
-                    }
-
-                    if (result != null)
-                    {
-                        results.Add(result);
-                    }
-                }
-            }
-
-            // return results.Count > 0? (mainFilePath, results): default;
-            return results;
-        }
-
-        private static SerializedInfo GetSerializedInfo(MemberInfo memberInfo, bool isProperty, Type targetType)
-        {
-            if (targetType == typeof(SaintsSerializedProperty))
-            {
-                return null;
-            }
-
-            Attribute[] attributes = ReflectCache.GetCustomAttributes<Attribute>(memberInfo);
-            bool hasNonSer = false;
-            bool hasSaintsSer = false;
-            foreach (Attribute attribute in attributes)
-            {
-                switch (attribute)
-                {
-                    case NonSerializedAttribute _:
-                        hasNonSer = true;
-                        break;
-                    case SaintsSerializedAttribute _:
-                        hasSaintsSer = true;
-                        break;
-                }
-            }
-
-            if (hasNonSer && !hasSaintsSer)
-            {
-                return null;
-            }
-
-            SaintsTargetCollection targetCollection = SaintsTargetCollection.FieldOrProperty;
-            Type elementType = null;
-            if (targetType.IsArray)
-            {
-                targetCollection = SaintsTargetCollection.Array;
-                elementType = targetType.GetElementType();
-            }
-            else
-            {
-                Type lisType = GetList(targetType);
-                // ReSharper disable once InvertIf
-                if (lisType != null)
-                {
-                    targetCollection = SaintsTargetCollection.List;
-                    elementType = lisType.GetGenericArguments()[0];
-                }
-            }
-
-            Type checkingType = targetCollection == SaintsTargetCollection.FieldOrProperty
-                ? targetType
-                : elementType;
-
-            Debug.Assert(checkingType != null, targetType);
-
-            // ReSharper disable once PossibleNullReferenceException
-            bool checkingIsEnum = checkingType.IsEnum;
-
-            // ReSharper disable once PossibleNullReferenceException
-            string fillTypeNameWithNameSpace = checkingType.FullName.Replace('+', '.');
-
-            if (!checkingIsEnum && IsGeneralClassOrStruct(checkingType))
-            {
-                if (!checkingType.IsDefined(typeof(SerializableAttribute)))
-                {
-                    return null;
-                }
-
-                SerializedInfo r = new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection, SaintsPropertyType.ClassOrStruct);
-
-                // class/struct check
-                foreach (MemberInfo subMember in checkingType
-                             .GetMembers(BindingFlags.Instance | BindingFlags.NonPublic |
-                                         BindingFlags.Public | BindingFlags.DeclaredOnly))
-                {
-                    SerializedInfo subSerInfo;
-                    switch (subMember)
-                    {
-                        case FieldInfo fi:
-                            subSerInfo = GetSerializedInfo(fi, false, fi.FieldType);
-                            break;
-                        case PropertyInfo pi:
-                            subSerInfo = GetSerializedInfo(pi, true, pi.PropertyType);
-                            break;
-                        default:
-                            continue;
-                    }
-
-                    if (subSerInfo != null)
-                    {
-                        r.SubFields.Add(subSerInfo);
-                    }
-                }
-
-                return r.SubFields.Count > 0 ? r : null;
-            }
-
-            if (!checkingIsEnum)
-            {
-                return null;
-            }
-
-            Type underType = checkingType.GetEnumUnderlyingType();
-            bool isLong = underType == typeof(long);
-            if (isLong)
-            {
-                return new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection,
-                    SaintsPropertyType.EnumLong);
-            }
-
-#if UNITY_2022_1_OR_NEWER
-            bool isULong = underType == typeof(ulong);
-            if (isULong)
-            {
-                return new SerializedInfo(memberInfo.Name, fillTypeNameWithNameSpace, isProperty, targetCollection,
-                    SaintsPropertyType.EnumULong);
-            }
-#endif
-
-            return null;
-        }
-
-        public static Type GetList(Type currentType)
-        {
-            if (!typeof(IEnumerable).IsAssignableFrom(currentType))
-            {
-                return null;
-            }
-
-            // ReSharper disable once LoopCanBeConvertedToQuery
-            foreach (Type baseType in ReflectUtils.GetGenBaseTypes(currentType))
-            {
-                if (baseType.GetGenericTypeDefinition() == typeof(List<>))
-                {
-                    return baseType;
-                }
-            }
-
-            return null;
-        }
-
-        private static bool IsGeneralClassOrStruct(Type type)
-        {
-            if (typeof(UnityEngine.Object).IsAssignableFrom(type))
-            {
-                return false;
-            }
-
-            if (type == typeof(SaintsSerializedProperty))
-            {
-                return false;
-            }
-
-            if (type.IsPrimitive || type == typeof(string) || type == typeof(decimal) || type.IsEnum)
-            {
-                return false;
-            }
-            return type.IsClass || (type.IsValueType && !type.IsPrimitive && !type.IsEnum);
-        }
+        // public static Type GetList(Type currentType)
+        // {
+        //     if (!typeof(IEnumerable).IsAssignableFrom(currentType))
+        //     {
+        //         return null;
+        //     }
+        //
+        //     // ReSharper disable once LoopCanBeConvertedToQuery
+        //     foreach (Type baseType in ReflectUtils.GetGenBaseTypes(currentType))
+        //     {
+        //         if (baseType.GetGenericTypeDefinition() == typeof(List<>))
+        //         {
+        //             return baseType;
+        //         }
+        //     }
+        //
+        //     return null;
+        // }
+        //
+        // private static bool IsGeneralClassOrStruct(Type type)
+        // {
+        //     if (typeof(UnityEngine.Object).IsAssignableFrom(type))
+        //     {
+        //         return false;
+        //     }
+        //
+        //     if (type == typeof(SaintsSerializedProperty))
+        //     {
+        //         return false;
+        //     }
+        //
+        //     if (type.IsPrimitive || type == typeof(string) || type == typeof(decimal) || type.IsEnum)
+        //     {
+        //         return false;
+        //     }
+        //     return type.IsClass || (type.IsValueType && !type.IsPrimitive && !type.IsEnum);
+        // }
     }
 }